Refactored everything, also Fixed list bugs where it would try to show 700 entries
