+++ test_data.json
@@ -1,0 +1,23 @@
+[
+    {
+        "brawlhalla_id": 6433263,
+        "name": "Zwergennest",
+        "rank": "Member",
+        "join_date": 1618333318,
+        "xp": 232800
+    },
+    {
+        "brawlhalla_id": 3764805,
+        "name": "Kyouno Natsumi <3",
+        "rank": "Officer",
+        "join_date": 1618497345,
+        "xp": 127110
+    },
+    {
+        "brawlhalla_id": 49936771,
+        "name": "Hunter",
+        "rank": "Officer",
+        "join_date": 1618949986,
+        "xp": 562464
+    }
+]
